+++ test/command/SendStartupMessage.ts
@@ -1,8 +1,7 @@
 import { SlackMessage } from "@atomist/slack-messages/SlackMessages";
-import { CommandHandler, Parameter, Secret } from "../../src/decorators";
-import { HandleCommand, HandlerContext, HandlerResult, Secrets } from "../../src/Handlers";
+import { CommandHandler, Parameter } from "../../src/decorators";
+import { HandleCommand, HandlerContext, HandlerResult } from "../../src/Handlers";
 import { sendMessages } from "../../src/operations/support/contextUtils";
-import { buttonForCommand, menuForCommand } from "../../src/spi/message/MessageClient";
 
 @CommandHandler("Sends a startup message to the owner of this automation-client")
 export class SendStartupMessage implements HandleCommand {