+++ test/internal/invoker/BuildableAutomationServerTest.ts
@@ -1,7 +1,10 @@
 import "mocha";
 import * as assert from "power-assert";
 
+import { SelfDescribingHandleCommand } from "../../../src/HandleCommand";
 import { consoleMessageClient } from "../../../src/internal/message/ConsoleMessageClient";
+import { succeed } from "../../../src/operations/support/contextUtils";
+import { AutomationServer } from "../../../src/server/AutomationServer";
 import { BuildableAutomationServer } from "../../../src/server/BuildableAutomationServer";
 import { HelloIssue } from "../../event/HelloIssue";
 import { AddAtomistSpringAgent, AlwaysOkEventHandler, FooBarEventHandler, FooBarIngestor } from "./TestHandlers";
@@ -11,18 +14,18 @@ const messageClient = consoleMessageClient;
 describe("BuildableAutomationServer", () => {
 
     it("should start with no rugs", () => {
-        const s = new BuildableAutomationServer({ name: "foobar", version: "1.0.0", teamIds: ["bar"], keywords: [] });
+        const s = new BuildableAutomationServer({name: "foobar", version: "1.0.0", teamIds: ["bar"], keywords: []});
         assert(s.rugs.commands.length === 0);
         assert(s.rugs.events.length === 0);
     });
 
     it("should register one no-arg handler and return its metadataFromInstance", () => {
-        const s = new BuildableAutomationServer({ name: "foobar", version: "1.0.0", teamIds: ["bar"], keywords: [] });
-        s.withCommandHandler(
-            { name: "foo", description: "foo", parameters: [], tags: [], intent: [], mapped_parameters: [] },
-            ch => Promise.resolve({
-                code: 0,
-            }));
+        const s = new BuildableAutomationServer({name: "foobar", version: "1.0.0", teamIds: ["bar"], keywords: []});
+        const h: SelfDescribingHandleCommand = {
+            name: "foo", description: "foo", parameters: [], tags: [], intent: [], mapped_parameters: [],
+            handle: succeed,
+        };
+        s.registerCommandHandler(() => h);
         assert(s.rugs.commands.length === 1);
         assert(s.rugs.events.length === 0);
         assert(s.rugs.commands[0].parameters.length === 0);
@@ -30,17 +33,15 @@ describe("BuildableAutomationServer", () => {
     });
 
     it("should register one single arg handler and return its metadataFromInstance", () => {
-        const s = new BuildableAutomationServer({ name: "foobar", version: "1.0.0", teamIds: ["bar"], keywords: [] });
-        s.withCommandHandler(
-            {
-                name: "foo", description: "foo", parameters: [{
+        const s = new BuildableAutomationServer({name: "foobar", version: "1.0.0", teamIds: ["bar"], keywords: []});
+        const h: SelfDescribingHandleCommand = {
+            name: "foo", description: "foo", parameters: [{
                 name: "one", description: "a thing", pattern: ".*", required: true,
             },
             ], tags: [], intent: [], mapped_parameters: [],
-            },
-            ch => Promise.resolve({
-                code: 0,
-            }));
+            handle: succeed,
+        };
+        s.registerCommandHandler(() => h);
         assert(s.rugs.commands.length === 1);
         assert(s.rugs.events.length === 0);
         assert(s.rugs.commands[0].parameters.length === 1);
@@ -48,17 +49,15 @@ describe("BuildableAutomationServer", () => {
     });
 
     it("should register one single arg handler and complain on invocation without parameter", () => {
-        const s = new BuildableAutomationServer({ name: "foobar", version: "1.0.0", teamIds: ["bar"], keywords: [] });
-        s.withCommandHandler(
-            {
-                name: "foo", description: "foo", parameters: [{
+        const s = new BuildableAutomationServer({name: "foobar", version: "1.0.0", teamIds: ["bar"], keywords: []});
+        const h: SelfDescribingHandleCommand = {
+            name: "foo", description: "foo", parameters: [{
                 name: "one", description: "a thing", pattern: ".*", required: true,
             },
             ], tags: [], intent: [], mapped_parameters: [],
-            },
-            ch => Promise.resolve({
-                code: 0,
-            }));
+            handle: succeed,
+        };
+        s.registerCommandHandler(() => h);
         assert.throws(() => {
             s.invokeCommand({
                 name: "foo",
@@ -72,17 +71,15 @@ describe("BuildableAutomationServer", () => {
     });
 
     it("should register one single arg handler and not complain on invocation without defaulted parameter", done => {
-        const s = new BuildableAutomationServer({ name: "foobar", version: "1.0.0", teamIds: ["bar"], keywords: [] });
-        s.withCommandHandler(
-            {
-                name: "foo", description: "foo", parameters: [{
+        const s = new BuildableAutomationServer({name: "foobar", version: "1.0.0", teamIds: ["bar"], keywords: []});
+        const h: SelfDescribingHandleCommand = {
+            name: "foo", description: "foo", parameters: [{
                 name: "one", description: "a thing", pattern: ".*", required: true, default_value: "banana",
             },
             ], tags: [], intent: [], mapped_parameters: [],
-            },
-            ch => Promise.resolve({
-                code: 0,
-            }));
+            handle: succeed,
+        };
+        s.registerCommandHandler(() => h);
         s.invokeCommand({
             name: "foo",
             args: [],
@@ -97,42 +94,44 @@ describe("BuildableAutomationServer", () => {
     });
 
     it("should register one single arg handler and invoke with valid parameter", done => {
-        const s = new BuildableAutomationServer({ name: "foobar", version: "1.0.0", teamIds: ["bar"], keywords: [] });
-        let paramVal: string;
-        s.withCommandHandler(
-            {
-                name: "foo", description: "foo", parameters: [{
+        const s = new BuildableAutomationServer({name: "foobar", version: "1.0.0", teamIds: ["bar"], keywords: []});
+        const h: SelfDescribingHandleCommand = {
+            name: "foo", description: "foo", parameters: [{
                 name: "one", description: "a thing", pattern: ".*", required: true,
             },
             ], tags: [], intent: [], mapped_parameters: [],
-            },
-            ch => {
-                paramVal = ch.args[0].value;
+            handle: (ch, params) => {
                 return Promise.resolve({
                     code: 0,
+                    paramVal: (params as any).one,
                 });
-            });
+            },
+        };
+        s.registerCommandHandler(() => h);
+        registerOneSingleArgHandlerAndInvokeWithValidParameter(s, done);
+    });
 
+    function registerOneSingleArgHandlerAndInvokeWithValidParameter(s: AutomationServer, done) {
         s.invokeCommand({
             name: "foo",
-            args: [{ name: "one", value: "value" }],
+            args: [{name: "one", value: "value"}],
         }, {
             teamId: "T666",
             correlationId: "555",
             messageClient,
-        }).then(_ => {
-            assert(paramVal === "value");
+        }).then(hr => {
+            assert((hr as any).paramVal === "value");
             done();
-        });
-    });
+        }).catch(done);
+    }
 
     it("should register one command handler instance and invoke with valid parameter", done => {
-        const s = new BuildableAutomationServer({ name: "foobar", version: "1.0.0", teamIds: ["bar"], keywords: [] });
-        s.fromCommandHandlerInstance(() => new AddAtomistSpringAgent());
+        const s = new BuildableAutomationServer({name: "foobar", version: "1.0.0", teamIds: ["bar"], keywords: []});
+        s.registerCommandHandler(AddAtomistSpringAgent);
         s.invokeCommand({
             name: "AddAtomistSpringAgent",
-            args: [{ name: "slackTeam", value: "T1691" }],
-            secrets: [{ name: "atomist://some_secret", value: "some_secret" }],
+            args: [{name: "slackTeam", value: "T1691"}],
+            secrets: [{name: "atomist://some_secret", value: "some_secret"}],
         }, {
             teamId: "T666",
             correlationId: "555",
@@ -142,7 +141,7 @@ describe("BuildableAutomationServer", () => {
     });
 
     it("should register one event handler instance and invoke with valid parameter", done => {
-        const s = new BuildableAutomationServer({ name: "foobar", version: "1.0.0", teamIds: ["bar"], keywords: [] });
+        const s = new BuildableAutomationServer({name: "foobar", version: "1.0.0", teamIds: ["bar"], keywords: []});
         s.fromEventHandlerInstance(() => new AlwaysOkEventHandler());
         s.onEvent({
             extensions: {
@@ -164,7 +163,7 @@ describe("BuildableAutomationServer", () => {
     });
 
     it("should register two event handler instances and invoke with valid parameter", done => {
-        const s = new BuildableAutomationServer({ name: "foobar", version: "1.0.0", teamIds: ["bar"], keywords: [] });
+        const s = new BuildableAutomationServer({name: "foobar", version: "1.0.0", teamIds: ["bar"], keywords: []});
         s.fromEventHandlerInstance(() => new AlwaysOkEventHandler());
         s.fromEventHandlerInstance(() => new FooBarEventHandler());
         s.onEvent({
@@ -188,7 +187,7 @@ describe("BuildableAutomationServer", () => {
     });
 
     it("should register one ingestor instance and invoke with valid payload", done => {
-        const s = new BuildableAutomationServer({ name: "foobar", version: "1.0.0", teamIds: ["bar"], keywords: [] });
+        const s = new BuildableAutomationServer({name: "foobar", version: "1.0.0", teamIds: ["bar"], keywords: []});
         s.fromIngestorInstance(() => new FooBarIngestor());
         s.onEvent({
             extensions: {
@@ -208,7 +207,7 @@ describe("BuildableAutomationServer", () => {
     });
 
     it("should register two ingestor instances and invoke with valid payload", done => {
-        const s = new BuildableAutomationServer({ name: "foobar", version: "1.0.0", teamIds: ["bar"], keywords: [] });
+        const s = new BuildableAutomationServer({name: "foobar", version: "1.0.0", teamIds: ["bar"], keywords: []});
         s.fromIngestorInstance(() => new FooBarIngestor());
         s.fromIngestorInstance(() => new HelloIssue());
         s.onEvent({
@@ -228,4 +227,5 @@ describe("BuildableAutomationServer", () => {
         });
     });
 
-});
+})
+;